+++ src/build.c
@@ -416,7 +416,7 @@ void editinput(void)
 {
 	char smooshyalign, repeatpanalign, *ptr, buffer[80];
 	short sectnum, nextsectnum, startwall, endwall, dasector, daang;
-	short mousx, mousy, mousz, bstatus;
+	long mousx, mousy, mousz, bstatus;
 	long i, j, k, cnt, templong=0, doubvel, changedir, wallfind[2], daz[2];
 	long dashade[2], goalz, xvect, yvect, hiz, loz;
 	short hitsect, hitwall, hitsprite;
@@ -2564,7 +2564,7 @@ void overheadeditor(void)
 	short pag, suckwall=0, sucksect, newnumwalls, newnumsectors, split=0, bad;
 	short splitsect=0, danumwalls, secondstartwall, joinsector[2], joinsectnum;
 	short splitstartwall=0, splitendwall, loopnum;
-	short mousx, mousy, bstatus;
+	long mousx, mousy, bstatus;
 	long centerx, centery, circlerad;
 	short circlewall, circlepoints, circleang1, circleang2, circleangdir;
 	long sectorhighlightx=0, sectorhighlighty=0;
